Allow only creators to edit their entities. Introduce new Moderator role

+++ src/main/resources/db/migration/sqlserver/V2.7.0.20190304162609__standardize-permissions-ir.sql
@@ -1,4 +1,8 @@
 delete from ${ohdsiSchema}.sec_role_permission where
+  permission_id in (select id from ${ohdsiSchema}.sec_permission where value in ('ir:*:delete'))
+  AND role_id = (SELECT id FROM ${ohdsiSchema}.sec_role WHERE name = 'Atlas users');
+
+delete from ${ohdsiSchema}.sec_role_permission where
   permission_id in (select id from ${ohdsiSchema}.sec_permission where value in ('ir:*:info'));
 
 delete from ${ohdsiSchema}.sec_permission where
